[RENAME] 파일의 대소문자 변경 밑작업 #NADA-49
