+++ src/pages/community/CommunityPage.js
@@ -8,10 +8,9 @@ import Filter from '../../containers/community/Filter';
 import { PostWriteButton } from '../../containers/community/PostWriteButton';
 import { setBarStatus } from '../../modules/bar';
 import { initializeAll } from '../../modules/filter';
-import { initializeAll as initializeAllPostWrite } from '../../modules/Community/postWrite';
+import { initializeAll as initializeAllPostWrite } from '../../modules/tommunity/postWrite';
 import {
   CommunityPage,
-  Search,
   Border,
   Contents,
 } from '../../styles/community/CommunityStyle';
